+++ projects/pastanaga-angular/src/lib/controls/toggles/toggles.module.ts
@@ -7,10 +7,11 @@ import { PaFormFieldModule } from '../form-field';
 import { ToggleComponent } from './toggle/toggle.component';
 import { RadioComponent } from './radio/radio.component';
 import { RadioGroupDirective } from './radio/radio-group.directive';
+import { PaTooltipModule } from '../../tooltip';
 
 @NgModule({
   declarations: [CheckboxComponent, ToggleComponent, RadioComponent, RadioGroupDirective],
-  imports: [CommonModule, FormsModule, ReactiveFormsModule, PaFocusableModule, PaFormFieldModule],
+  imports: [CommonModule, FormsModule, ReactiveFormsModule, PaFocusableModule, PaFormFieldModule, PaTooltipModule],
   exports: [CheckboxComponent, ToggleComponent, RadioComponent, RadioGroupDirective],
 })
 export class PaTogglesModule {}